DATS-000 feat: dev page
