+++ src/components/GraphBox.tsx
@@ -1,7 +1,14 @@
-import type { HTMLAttributes, ReactNode } from 'react';
+import type { HTMLAttributes } from 'react';
 import React from 'react';
 import classnames from 'classnames';
 
+/**
+ *
+ * @param title
+ * @param children
+ * @param className
+ * @constructor
+ */
 export const GraphBox = ({
   title,
   children,
@@ -9,7 +16,7 @@ export const GraphBox = ({
 }: HTMLAttributes<HTMLDivElement> & {
   title: string;
 }) => (
-  <div className={classnames('bg-white rounded-lg shadow max-w-max p-6', className)}>
+  <div className={classnames('bg-white rounded-lg shadow max-w-max p-3', className)}>
     <div className="font-medium pb-4">{title}</div>
     {children}
   </div>
